1. Add AAZResourceLocationArg and AAZResourceIdArg. 2. rename AAZContentArgBrowser by AAZArgBrower. 3. Add format_args function. To support argument format in future.
