+++ src/azure-cli-core/azure/cli/core/aaz/__init__.py
@@ -1,6 +1,6 @@
 from ._base import AAZValuePatch, AAZUndefined
-from ._arg import AAZArgumentsSchema, AAZArgEnum, AAZStrArg, AAZIntArg, AAZObjectArg, AAZDictArg, AAZFloatArg, \
-    AAZBaseArg, AAZBoolArg, AAZListArg, AAZResourceGroupNameArg, has_value
+from ._arg import has_value, AAZArgumentsSchema, AAZArgEnum, AAZStrArg, AAZIntArg, AAZObjectArg, AAZDictArg, AAZFloatArg, \
+    AAZBaseArg, AAZBoolArg, AAZListArg, AAZResourceGroupNameArg, AAZResourceLocationArg, AAZResourceIdArg
 from ._command import AAZCommand, AAZCommandGroup, register_command, register_command_group, load_aaz_command_table
 from ._field_type import AAZIntType, AAZFloatType, AAZStrType, AAZBoolType, AAZDictType, AAZListType, AAZObjectType
 from ._operation import AAZHttpOperation, AAZInstanceUpdateOperation
